requirements.txt
attrs==20.3.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
flake8==3.9.0
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
google-api-core==1.26.3
google-auth==1.30.0
google-cloud-secret-manager==2.4.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpc-google-iam-v1==0.12.3
grpcio==1.37.1
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
libcst==0.3.18
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.9
pluggy==0.13.1
proto-plus==1.18.1
protobuf==3.16.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.13
toml==0.10.2
typing-extensions==3.10.0.0
typing-inspect==0.6.0
urllib3==1.26.4
uWSGI==2.0.19.1
Werkzeug==1.0.1
